Move common code to a new dir (#682)
