+++ coreblocks/arch/__init__.py
@@ -1,0 +1,3 @@
+from .optypes import *  # noqa: F401
+from .isa import *  # noqa: F401
+from .isa_consts import *  # noqa: F401
