--- coreblocks/frontend/decoder/__init__.py
@@ -1,2 +1,0 @@
-from .isa import *  # noqa: F401
-from .optypes import *  # noqa: F401
