+++ coreblocks/frontend/decoder/instr_description.py
@@ -1,9 +1,7 @@
 from dataclasses import KW_ONLY, dataclass
 from typing import Optional
 
-from coreblocks.params import *
-from .isa import *
-from .optypes import *
+from coreblocks.arch import *
 
 
 @dataclass(frozen=True)
