+++ coreblocks/params/__init__.py
@@ -1,4 +1,3 @@
-from .isa_params import *  # noqa: F401
 from .genparams import *  # noqa: F401
 from .fu_params import *  # noqa: F401
 from .icache_params import *  # noqa: F401
