+++ coreblocks/params/fu_params.py
@@ -3,15 +3,14 @@ from dataclasses import dataclass
 from collections.abc import Collection, Iterable
 
 from coreblocks.func_blocks.interface.func_protocols import FuncBlock, FuncUnit
-from coreblocks.params.isa_params import Extension, extension_implications
-from coreblocks.frontend.decoder import optypes_required_by_extensions
+from coreblocks.arch.isa import Extension, extension_implications
+from coreblocks.arch.optypes import optypes_required_by_extensions, OpType
 
 from typing import TYPE_CHECKING
 
 
 if TYPE_CHECKING:
     from coreblocks.params.genparams import GenParams
-    from coreblocks.frontend.decoder.optypes import OpType
 
 
 __all__ = [
