+++ test/func_blocks/fu/test_mul_unit.py
@@ -1,6 +1,6 @@
 from parameterized import parameterized_class
 
-from coreblocks.frontend.decoder import Funct3, Funct7, OpType
+from coreblocks.arch import Funct3, Funct7, OpType
 from coreblocks.func_blocks.fu.mul_unit import MulFn, MulComponent, MulType
 
 from transactron.utils import signed_to_int, int_to_signed
