+++ test/params/test_configurations.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 
 from coreblocks.params.genparams import GenParams
 from coreblocks.params.configurations import *
-from coreblocks.params.isa_params import gen_isa_string
+from coreblocks.arch.isa import gen_isa_string
 from coreblocks.params.fu_params import extensions_supported
